+++ src/Main.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import App from "./App";
-import { MainPage } from "./page-layout/MainPage/MainPage";
-import { TestPage } from "./page-layout/TestPage/TestPage";
+import { MainPage } from "./pages/MainPage/MainPage"
+import { TestPage } from "./pages/TestPage/TestPage";
 
 function Main() {
   return (
