+++ src/components/feature/feature.txt
@@ -1,0 +1,2 @@
+파일을 넣지 않으면 폴더가 깃허브에 올라가지 않아서 만들었습니다.
+나중에 feature 폴더에 파일을 넣으시면 이건 삭제해주세요.
