+++ src/components/ui/InputField/InputField.module.css
@@ -1,0 +1,26 @@
+.input-wrap {
+  position: relative;
+  max-width: 336px;
+}
+.input {
+  width: 100%;
+  padding: 12px 16px 12px 42px;
+  border: 1px solid #818181;
+  border-radius: 8px;
+  outline: none;
+  font-size: 16px;
+  line-height: 22px;
+  box-sizing: border-box;
+}
+.input:focus {
+  border: 1px solid #542f1a;
+}
+.icon {
+  position: absolute;
+  top: 50%;
+  left: 16px;
+  width: 20px;
+  height: 20px;
+  transform: translateY(-50%);
+  fill: #818181;
+}
