--- src/page-layout/MainPage/index.js
@@ -1,1 +1,0 @@
-export * from "./MainPage";
