--- src/page-layout/TestPage/index.js
@@ -1,1 +1,0 @@
-export * from "./TestPage";
