+++ src/pages/MainPage/index.js
@@ -1,0 +1,1 @@
+export * from "./MainPage";
