+++ src/styles/reset.css
@@ -1,0 +1,75 @@
+/* user agent stylesheet 초기화 */
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  font-family: "Pretendard";
+  word-break: keep-all;
+  color: var(--grayscale-60, #000);
+  font-style: normal;
+  line-height: normal;
+}
+
+h1 {
+  font-size: 40px;
+  font-weight: 400;
+}
+h2 {
+  font-size: 32px;
+  font-weight: 400;
+  line-height: 40px;
+}
+h3 {
+  font-size: 24px;
+  font-weight: 400;
+  line-height: 30px;
+}
+
+.body-1-regular{
+  font-size: 20px;
+  font-weight: 400;
+  line-height: 25px;
+}
+.body-1-bold{
+  font-size: 20px;
+  font-weight: 400;
+  line-height: 25px;
+}
+
+.body-2-regular{
+  font-size: 18px;
+  font-weight: 400;
+  line-height: 24px;
+}
+.body-2-bold {
+  font-size: 18px;
+  font-weight: 400;
+  line-height: 24px;
+}
+
+.body-3-regular{
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 22px;
+}
+.body-3-bold {
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 22px;
+}
+
+.caption-1-regular{
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 18px;
+}
+.caption-1-medium{
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 18px;
+}
+.caption-1-bold{
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 18px;
+}
