--- src/ui/InputTextarea/InputTextarea.module.css
@@ -1,25 +1,0 @@
-.inputarea {
-  width: 336px;
-  height: 140px;
-}
-
-.input {
-  display: flex;
-  padding: 16px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  border-radius: 8px;
-  background: var(--grayscale-20);
-}
-
-.input:focus,
-textarea:focus {
-  border: 1px solid #ff0000; /* 빨간색으로 변경 */
-}
-
-.text {
-  color: var(--grayscale-40, #818181);
-  font-feature-settings: "clig" off, "liga" off;
-  line-height: 22px; /* 137.5% */
-}
