--- src/variables.css
@@ -1,9 +1,0 @@
-.shadow-1 {
-  box-shadow: 0px 4px 4px 0px rgba(140, 140, 140, 0.25);
-}
-.shadow-2 {
-  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-}
-.shadow-3 {
-  box-shadow: 0px 16px 20px 0px rgba(48, 48, 48, 0.62);
-}
